+++ rs/execution_environment/src/hypervisor/tests.rs
@@ -7141,40 +7141,6 @@ fn yield_for_dirty_page_copy_does_not_trigger_on_system_subnets_without_dts() {
 }
 
 #[test]
-fn table_grow_fails_with_too_many_elements() {
-    let wat = r#"
-        (module
-            (import "ic0" "msg_reply" (func $msg_reply))
-            (import "ic0" "msg_reply_data_append"
-                (func $msg_reply_data_append (param i32 i32))
-            )
-            (func (export "canister_update table_grow")
-                (i32.store (i32.const 0)
-                    (table.grow 0 (ref.null extern) (i32.const 2000000))
-                )
-                (call $msg_reply_data_append (i32.const 0) (i32.const 4))
-                (call $msg_reply)
-            )
-            (table 0 externref)
-            (memory 1)
-        )
-    "#;
-    let mut test = ExecutionTestBuilder::new().build();
-    let canister_id = test.canister_from_wat(wat).unwrap();
-    let ingress_id = test.ingress_raw(canister_id, "table_grow", vec![]).0;
-    test.execute_message(canister_id);
-    let ingress_status = test.ingress_status(&ingress_id);
-    let ingress_state = match ingress_status {
-        IngressStatus::Known { state, .. } => state,
-        IngressStatus::Unknown => unreachable!("Expected known ingress status"),
-    };
-    assert_eq!(
-        IngressState::Completed(WasmResult::Reply((-1_i32).to_le_bytes().to_vec())),
-        ingress_state
-    );
-}
-
-#[test]
 fn declaring_too_many_tables_fails() {
     let wat = format!("(module {})", "(table 0 externref)".repeat(100));
     let mut test = ExecutionTestBuilder::new().build();
@@ -7249,7 +7215,7 @@ fn wasm_memory_limit_is_enforced_at_start_of_update() {
         .unwrap();
     assert_eq!(WasmResult::Reply(100_i32.to_le_bytes().to_vec()), result);
 
-    test.canister_update_wasm_memory_limit(canister_id, NumBytes::new(0))
+    test.canister_update_wasm_memory_limit(canister_id, NumBytes::new(1))
         .unwrap();
 
     let err = test
@@ -7261,6 +7227,42 @@ fn wasm_memory_limit_is_enforced_at_start_of_update() {
         .unwrap_err();
 
     assert_eq!(err.code(), ErrorCode::CanisterWasmMemoryLimitExceeded);
+}
+
+#[test]
+fn wasm_memory_limit_zero_means_unlimited() {
+    let mut test = ExecutionTestBuilder::new().build();
+    let canister_id = test.universal_canister().unwrap();
+
+    let result = test
+        .ingress(canister_id, "update", use_wasm_memory_and_reply(100))
+        .unwrap();
+    assert_eq!(WasmResult::Reply(100_i32.to_le_bytes().to_vec()), result);
+
+    // The execution fails.
+    test.canister_update_wasm_memory_limit(canister_id, NumBytes::new(1))
+        .unwrap();
+
+    let err = test
+        .ingress(
+            canister_id,
+            "update",
+            wasm().push_bytes(&[]).append_and_reply().build(),
+        )
+        .unwrap_err();
+
+    assert_eq!(err.code(), ErrorCode::CanisterWasmMemoryLimitExceeded);
+
+    test.canister_update_wasm_memory_limit(canister_id, NumBytes::new(0))
+        .unwrap();
+
+    // The execution succeeds.
+    test.ingress(
+        canister_id,
+        "update",
+        wasm().push_bytes(&[]).append_and_reply().build(),
+    )
+    .unwrap();
 }
 
 #[test]